Tornado fx

+++ src/main/kotlin/models/MaterialModel.kt
@@ -1,9 +1,10 @@
 package models
 
-data class MaterialModel (var materialId: Long = 0,
-                          var materialName: String = "",
-                          var materialType: String = "",
-                          var materialWeight: Int = 0,
-                          var materialPrice: Double = 0.0)
+data class MaterialModel(
+    var materialId: Long = 0,
+    var materialName: String = "",
+    var materialType: String = "",
+    var materialWeight: Int = 0,
+    var materialPrice: Float = 0.0f)
 {
 }
